# packages/shared/openapi.yaml
openapi: 3.0.3
info:
  title: XinChat API
  version: 0.3.0
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
paths:
...
  /v1/friends:
    get:
      security: [{ bearerAuth: [] }]
      summary: List friends and requests
      responses:
        "200":
          description: Friends
...
  /v1/friends/{id}:
    patch:
      security: [{ bearerAuth: [] }]
      summary: Update friend note/tags
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
...